separate move kmeans out
