+++ rust/Cargo.toml
@@ -33,7 +33,9 @@ half = { "version" = "=2.2.1", default-features = false, features = [
 ] }
 approx = "0.5.1"
 criterion = { version = "0.5", features = ["async", "async_tokio"] }
+futures = "0.3"
 log = "0.4"
+num_cpus = "1.0"
 num-traits = "0.2"
 object_store = { version = "0.6.1", features = ["aws", "gcp", "azure"] }
 pprof = { version = "0.12", features = ["flamegraph", "criterion"] }
@@ -41,6 +43,7 @@ rand = { version = "0.8.3", features = ["small_rng"] }
 serde = { version = "^1" }
 serde_json = { version = "1" }
 snafu = "0.7.4"
+tokio = { version = "1.23", features = ["rt-multi-thread"] }
 
 [profile.bench]
 opt-level = 3
