+++ rust/lance-linalg/Cargo.toml
@@ -8,9 +8,14 @@ repository = { workspace = true }
 [dependencies]
 arrow-array = { workspace = true }
 arrow-schema = { workspace = true }
+arrow-select = { workspace = true }
+futures = { workspace = true }
 half = { workspace = true }
+log = { workspace = true }
+num_cpus = { workspace = true }
 num-traits = { workspace = true }
 rand = { workspace = true }
+tokio = { workspace = true }
 
 [dev-dependencies]
 approx = { workspace = true }
