+++ rust/lance/Cargo.toml
@@ -59,10 +59,10 @@ pin-project = "1.0"
 prost = "0.10"
 prost-types = "0.10"
 roaring = "0.10.1"
-tokio = { version = "1.23", features = ["rt-multi-thread"] }
+tokio = { workspace = true }
 url = "2.3"
 rand = { version = "0.8.3", features = ["small_rng"] }
-futures = "0.3.27"
+futures = { workspace = true }
 uuid = { version = "1.2", features = ["v4"] }
 path-absolutize = "3.0.14"
 shellexpand = "3.0.0"
