+++ rust/lance/src/dataset/scanner.rs
@@ -34,13 +34,14 @@ use datafusion::physical_plan::{
 use datafusion::prelude::*;
 use datafusion::scalar::ScalarValue;
 use futures::stream::{Stream, StreamExt};
+use lance_linalg::distance::MetricType;
 use log::debug;
 
 use super::Dataset;
 use crate::datafusion::physical_expr::column_names_in_expr;
 use crate::datatypes::Schema;
 use crate::format::{Fragment, Index};
-use crate::index::vector::{MetricType, Query};
+use crate::index::vector::Query;
 use crate::io::exec::{
     KNNFlatExec, KNNIndexExec, LanceScanExec, Planner, ProjectionExec, TakeExec,
 };
